requirement.txt
torch==2.3.1
torchvision==0.18.1
transformers>=4.45.0
litserve==0.2.3
accelerate==1.0.0
fastapi==0.115.0
fire==0.7.0
google-cloud-aiplatform==1.70.0
google-generativeai==0.8.3
gunicorn==23.0.0
loguru==0.7.2
mesop==0.12.9
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
pyarrow==17.0.0
pydantic==2.7.4
srsly==2.4.8
tqdm==4.66.5
weaviate-client==4.8.1
sentencepiece
pyyaml
ninja
einops
psutil
packaging
ruff
pre-commit
gpustat